refactor(monitoring): extract ThreadsFiltersPopover into its own file
threads.tsx (1037 lines) is one of the largest frontend files. The
filters popover is a self-contained unit — it only reads its own props
and local `open` state, sharing no closures with the rest of the file —
and was already imported independently by monitoring/index.tsx, so this
is a byte-identical move plus updated imports, not a rewrite.
